+++ components/FolderButton/FolderButton.styled.ts
@@ -16,8 +16,8 @@ export const FolderName = styled.span<{ $select: string | boolean }>`
   color: ${(props) => (props.$select === 'select' ? '#fff' : '#000')};
 
   &:hover {
-    background-color: var(--Linkbrary-gray10);
-    color: #000;
+    background-color: var(--Primary);
+    color: #fff;
   }
 
   @media (max-width: 768px) {
